# mvps-api/src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  title: Echo Healthcare API
  description: |
    Multi-Vendor Healthcare Products Management System API
    
    This API provides comprehensive endpoints for managing vendors, products, pricing, 
    inventory, orders, and customers for Echo Healthcare's online and door-to-door sales platform.
    
    Key Features:
    - Automatic lowest price vendor selection
    - Multi-vendor product management
    - Real-time inventory tracking
    - Order management (online & door-to-door)
    - Customer management
    - Sales analytics and reporting
  version: 1.0.0
  contact:
    name: Echo Healthcare API Support
    email: support@echo-healthcare.com
servers:
  - url: http://localhost:8888/v1
    description: Development server
...
paths:
...
  /vendors/{vendorId}/products:
    get:
      tags:
        - Vendors
      summary: Get all products from a vendor
      description: Retrieve all products offered by a specific vendor
      parameters:
        - $ref: '#/components/parameters/VendorId'
        - name: is_available
          in: query
          description: Filter by availability
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VendorProduct'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    VendorId:
      name: vendorId
      in: path
      required: true
      description: Vendor ID
      schema:
        type: integer
...
  schemas:
...
    VendorProduct:
      type: object
      properties:
        vendor_product_id:
          type: integer
          example: 1
        vendor_id:
          type: integer
          example: 1
        vendor_name:
          type: string
          example: "Pharma Solutions"
          description: Name of the vendor (read-only, populated from vendor record)
        product_id:
          type: integer
          example: 1
        vendor_sku:
          type: string
          example: V1-PAR500
        cost_price:
          type: number
          format: decimal
          example: 25.00
        mrp:
          type: number
          format: decimal
          example: 30.00
        discount_percentage:
          type: number
          format: decimal
          example: 5.00
        final_price:
          type: number
          format: decimal
          example: 23.75
          description: Auto-calculated as cost_price * (1 - discount_percentage/100)
        minimum_order_quantity:
          type: integer
          example: 1
        stock_quantity:
          type: integer
          example: 500
        expiry_date:
          type: string
          format: date
          nullable: true
          example: "2026-12-31"
        is_available:
          type: boolean
          example: true
        delivery_time_days:
          type: integer
          example: 7
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authorization header using the Bearer scheme
security:
  - BearerAuth: []